fix(supabase): keep scene_num as text in sync_shooting_days RPC
scenes.scene_number is a TEXT column, not bigint. The original
migration cast `(elem->>'n')::bigint` and the comparison
`s.scene_number = a.scene_num` errored with
`operator does not exist: text = bigint` on apply. Even with the
right cast it would have broken on non-numeric placeholders some
screenplays use ("P1", "TBC", etc).

Drop the cast — keep the value as text the whole way through. The
JS helper already types `n: string` (parseSceneRef returns string),
and the doc comments on both sides now spell that contract out.

Audit habit going forward: for migrations that touch existing
tables, verify column types against the live schema (information_
schema.columns or \\d) before writing the function body. The audit
on the original prompt validated the helper logic but not the
schema match.

+++ supabase/migrations/025_sync_shooting_days_rpc.sql
@@ -15,10 +15,13 @@
 --   [{ "n": "176", "s": "A",  "d": 13 },
 --    { "n": "4",   "s": null, "d": 1  }]
 --
--- where n = scene_number (string, cast to bigint), s = number_suffix
--- (nullable), d = shooting_day. The match is project-scoped and
--- NULL-safe via `IS NOT DISTINCT FROM` so the same SQL handles both
--- suffixed and unsuffixed rows.
+-- where n = scene_number (text — `scenes.scene_number` is a TEXT
+-- column, so we keep the value as text and compare text-to-text;
+-- this also lets legitimate non-numeric scene numbers like "P1" or
+-- "TBC" round-trip cleanly), s = number_suffix (nullable),
+-- d = shooting_day. The match is project-scoped and NULL-safe via
+-- `IS NOT DISTINCT FROM` so the same SQL handles both suffixed and
+-- unsuffixed rows.
 --
 -- Returns the number of rows actually updated so the caller can
 -- compare it against the input length and log unmatched references.
@@ -39,7 +42,7 @@ DECLARE
 BEGIN
   WITH a AS (
     SELECT
-      (elem->>'n')::bigint    AS scene_num,
+      (elem->>'n')            AS scene_num,
       NULLIF(elem->>'s', '')  AS suffix,
       (elem->>'d')::integer   AS shooting_day
     FROM jsonb_array_elements(p_assignments) AS elem
